Milieu- en habitatpreferenties van Nederlandse zoetwatermacrofauna

Deze dataset bevat autecologische informatie van Nederlandse macrofaunasoorten en is opgesteld door de subgroep Autecologie van de Werkgroep Ecologisch Waterbeheer.

wew_mafa_data

Format

Dataframe

taxonnaam_orig Taxonnaam uit originele dataset

taxonnaam_voorkeur Voorkeursnaam taxon. Als deze ontbreekt betekent dat dat de taxonnaam_orig niet in de TWN-lijst voorkomt.

indicator_groep Het type milieufactor

indicator_waarde Waarde van de milieufactor

indicator_waarde_aandeel Relatief aandeel van de waarde binnen de milieufactor op een schaal van 0 tot 10.

Source

http://www.wew.nu/publicaties/wew_themanummer23_soortenlijst.xls

http://www.wew.nu/publicaties/wew_themanummer23.pdf

Details

Voor negen factoren zijn de milieu- en habitatpreferenties van macrofaunasoorten verzameld, namelijk zoutgehalte, zuurgraad, droogval, stroomsnelheid, substraat, diepte, oppervlakte, saprobie en trofie. Deze factoren zijn gekozen omdat ze sterk sturend zijn en omdat er voldoende gegevens van beschikbaar zijn. Voor elke factor zijn vier tot tien klassen onderscheiden.

Metadata dataset

Naam: Milieu- en habitatpreferenties van Nederlandse zoetwatermacrofauna

Bronhouder: Werkgroep Ecologisch Waterbeheer

Auteurs:

Auteurs: W.C.E.P. Verberk, P.F.M. Verdonschot, T. van Haaren, B. van Maanen.

Redactie: Rebi C. Nijboer.

Subgroep Autoecologie, WEW-23.

Datum dataset: Mei 2012.

Datum opname dataset: 10 April 2019

Gebruiksrechten: Geen beperkingen

Soortgroep: Macrofauna

Bewerkingen

Ten behoeve van eenvoudiger gebruik van de dataset is de dataset omgezet in een lang formaat. Ontbrekende waarden zijn niet opgenomen.

De zeldzaamheid is opgenomen in de aparte tabel wew_mafa_zeldzaamheid.

Een toelichting op de indicatoren is opgenomen in wew_mafa_toelichting.

See also

Examples

head(wew_mafa_data)

#> # A tibble: 6 x 5
#>   taxonnaam_orig taxonnaam_voork~ indicator_groep indicator_waarde
#>   <chr>          <chr>            <chr>           <chr>           
#> 1 Ablabesmyia l~ Ablabesmyia lon~ zoutgehalte     zoet            
#> 2 Ablabesmyia m~ Ablabesmyia mon~ zoutgehalte     zoet            
#> 3 Ablabesmyia p~ Ablabesmyia pha~ zoutgehalte     zoet            
#> 4 Acamptocladiu~ Acamptocladius ~ zoutgehalte     zoet            
#> 5 Acentria ephe~ Acentria epheme~ zoutgehalte     zoet            
#> 6 Acilius canal~ Acilius canalic~ zoutgehalte     zoet            
#> # ... with 1 more variable: indicator_waarde_aandeel <dbl>